+++ packages/shared/src/server/landing-page/landing-page.tsx
@@ -2,11 +2,32 @@ import * as React from "react";
 import styles from "./landing-page.module.scss";
 import { Cards } from "../cards";
 import { Logo } from "../logo";
+import featured from "../../../../../featured.json";
 
 interface LandingPageProps {
   title: string;
   children?: React.ReactNode;
 }
+
+const cards = [
+  {
+    href: "https://react18-tools.github.io/turborepo-template/",
+    title: "Docs",
+    description: "Check out the official documentation for more information.",
+  },
+  {
+    href: "https://github.com/react18-tools/turborepo-template",
+    title: "More Examples",
+    description:
+      "Check out more examples on the official GitHub Repo. Feel free to suggest additional examples in the discussions section.",
+  },
+  {
+    href: "https://github.com/react18-tools/turborepo-template",
+    title: "Star this repo",
+    description:
+      "Star this repo for your new library! This also motivates us and helps us understand that community is interested in this work.",
+  },
+];
 
 /**
  * # LandingPage
@@ -28,7 +49,12 @@ export function LandingPage({ title, children }: LandingPageProps) {
         </p>
       </div>
       {children}
-      <Cards />
+      <div className={styles.featured}>
+        <h2>Featured packages built with this template.</h2>
+        <hr />
+        <Cards cards={featured} />
+      </div>
+      <Cards cards={cards} />
     </main>
   );
 }
